refactor(ui): consolidate the dropdown system into one component (#203)
Following the initial themed-surface work, collapse the dropdown family into a
single cohesive unit and remove the duplication it carried:

- Merge NimazDropdownField + NimazDropdownMenu + NimazDropdownMenuItem (three
  files) into one NimazDropdown.kt.
- Replace the two near-identical row composables (NimazDropdownMenuItem for
  selection, NimazDropdownAction for actions) with ONE NimazDropdownRow: pass
  `selected` for a value choice (accent fill + circular check) or `destructive`
  for an irreversible command.
- NimazDropdownField now builds on NimazDropdownMenu + NimazDropdownRow instead
  of inlining its own Material DropdownMenu call, so there is a single anchored-
  menu surface code path.
- Update the Qaida overflow call site (NimazDropdownAction -> NimazDropdownRow).
- Merge the three dropdown test files into one NimazDropdownTest (9 cases).
- Update the ARCHITECTURE.md decision rule for the single-file system.

Same package, so the 4 NimazDropdownField call sites need no changes.
:app:compileDebugKotlin + NimazDropdownTest green.
